Add test-case for passing NaN argument to sumAll
Both the requirements from the provided tests and the provided solution
suggest that passing a non-number argument should return 'ERROR'.
The current tests however, allow for an implementation using 'typeof'
which wouldn't fail the current test cases for 'non-number parameters'
while at the same time not prohibiting NaN which is the epitome of a
not-a-number instance that can cause logical errors.

The addition of this test aims to prevent such solutions and remind
learners the pitfalls of using typeof when checking for numbers, like
the number value NaN.